Add a skeleton for the cancel endpoint, need to discuss SoC and what to name the cancel method in the service layer
